refactor(memory): move host leaves out of implementation
