+++ core/memory/module.py
@@ -32,7 +32,7 @@ from core.memory.everos import (
     MemoryProviderFailure,
     MemoryProviderPort,
 )
-from core.memory.project_ids import (
+from vibe.memory_project_ids import (
     is_new_stored_memory_project_id,
     is_persisted_memory_project_id,
     is_writable_memory_project_id,
