+++ core/memory/processing_record.py
@@ -37,10 +37,6 @@ PROCESSING_RECORD_WORK_TIMEOUT_SECONDS = (
         * PROCESSING_RECORD_SQLITE_BOUND_SECONDS,
     )
 )
-PROCESSING_RECORD_TRANSPORT_TIMEOUT_SECONDS = (
-    PROCESSING_RECORD_WORK_TIMEOUT_SECONDS
-    + PROCESSING_RECORD_TRANSPORT_MARGIN_SECONDS
-)
 
 
 _AwaitResult = TypeVar("_AwaitResult")
